Generate note link in DataStore
Generate the note link in the DataStore so it will only need to be done
once. In the reporter we can use the link but now we need to extract the
note number for display purposes. To do this we use a regex, it has the
benefit over using slice `slice` that it will keep working when there is
a note with double digits.