Merge polyfill links and show in report
The DataStore will now merge links from the additional data source with
the Can I Use dataset when it detects the feature key has already been
processed. Whatever links are present in the additional data source will
be added to those from Can I Use. There is an attempt to filter out
duplicate URLs.

The Reporter will show the links to the polyfills after showing the
browser support for the category and before the notes.