Change CodeInput check button behaviour
If someone clicked on the button to check the compatibility without
providing code in the text area and without selecting a file first there
would just be a line in the log mentioning that the text area is empty.

To make this more user friendly CodeInput will now show the file dialog
when the visitor clicks on the button when no code or file has been
provided. When the file dialog is launched in this fashion the selected
file will automatically be loaded and a report is generated without
having to press the button a second time.