Make browser support testing more robust
The browser support was done by taking the first char from the string.
This has been changed to using a regex to get a support character that
we support. This will make the code better when a datasource would list
a modifier as 'd' before a support indicator like 'a'.