Fix ScollTo behaviour
Because ScrollTo has been given a correction option there was an issue
with the element which allows the visitor to scroll to the top. The
correction was also applied to this action. The module now recognizes
the attribute `data-no-correction`. When this attribute is present the
correction will not applied to the destination.