Refactor filtering
With the addition of a second filter block it made sense to refactor how
the filters are setup. This involved creating a base class for all sorts
of filters. As a result the original BrowserFilter had to be refactored
to use this new base class.

There is now a new filter class called SupportFilter. Through this the
visitor can manage which support categories should be shown in the
report.

The reporter class now has two methods for filtering, one for the
browsers and one for the support categories. Both call the same internal
method to handle the filtering of the elements.