Refactor compatibility data loading
As a first step to make the application logic usable for other platforms
like Node.js the loading of the compatibility data has been moved to its
own module. This will make it easier to make a version where the data is
loaded from the file system instead of an URL.

All logic concerning the loading of compatibility data has been removed
from CodeAnalyzer and DataStore and moved to App and AjaxLoader.

In addition the App module has been changed to delay the creation of
some modules just before they're needed instead of on App initialization.