Add optional filename param to CodeAnalyzer
The CodeAnalyzer `check` method has been given an optional `fileName`
parameter. This can be used when analysing multiple files at a time to
keep track of which file a particular instance of CodeAnalyzer was
working on.

The result message posted by CodeAnalyzer has been expanded with a
`sender` property. This is the instance of CodeAnalyzer which has posted
the message to the namespace. This way the receiver of the `codeAnalyzed`
message has a way of knowing which instance of CodeAnalyzer is done.